requirements.txt
# for processing
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

hdbscan
nltk
pandas
pyarrow
transformers
umap-learn

# for noteboks
ipykernel
ipython
jupyterlab
matplotlib
seaborn
plotly
nbformat
